+++ src/mpd_shared/mpd_shared_sticker.c
@@ -68,19 +68,19 @@ bool mpd_shared_get_sticker(struct t_mpd_state *mpd_state, const char *uri, stru
 
     while ((pair = mpd_recv_sticker(mpd_state->conn)) != NULL) {
         if (strcmp(pair->name, "playCount") == 0) {
-            sticker->playCount = strtoimax(pair->value, &crap, 10);
+            sticker->playCount = (int)strtoimax(pair->value, &crap, 10);
         }
         else if (strcmp(pair->name, "skipCount") == 0) {
-            sticker->skipCount = strtoimax(pair->value, &crap, 10);
+            sticker->skipCount = (int)strtoimax(pair->value, &crap, 10);
         }
         else if (strcmp(pair->name, "lastPlayed") == 0) {
-            sticker->lastPlayed = strtoimax(pair->value, &crap, 10);
+            sticker->lastPlayed = (int)strtoimax(pair->value, &crap, 10);
         }
         else if (strcmp(pair->name, "lastSkipped") == 0) {
-            sticker->lastSkipped = strtoimax(pair->value, &crap, 10);
+            sticker->lastSkipped = (int)strtoimax(pair->value, &crap, 10);
         }
         else if (strcmp(pair->name, "like") == 0) {
-            sticker->like = strtoimax(pair->value, &crap, 10);
+            sticker->like = (int)strtoimax(pair->value, &crap, 10);
         }
         mpd_return_sticker(mpd_state->conn, pair);
     }
