+++ src/mympd_api.c
@@ -106,7 +106,7 @@ void mympd_autoconf(struct t_mympd_state *mympd_state) {
         sdsfree(mpd_pass);
         
         sds mpd_port = get_mpd_conf("port", mympd_state->mpd_state->mpd_host);
-        mympd_state->mpd_state->mpd_port = strtoimax(mpd_port, NULL, 10);
+        mympd_state->mpd_state->mpd_port = (int)strtoimax(mpd_port, NULL, 10);
         sdsfree(mpd_port);
         
         sds music_directory = get_mpd_conf("music_directory", mympd_state->music_directory);
@@ -124,7 +124,7 @@ void mympd_autoconf(struct t_mympd_state *mympd_state) {
         if (mpd_host != NULL && strlen(mpd_host) <= 100) {
             if (mpd_host[0] != '@' && strstr(mpd_host, "@") != NULL) {
                 int count;
-                sds *tokens = sdssplitlen(mpd_host, strlen(mpd_host), "@", 1, &count);
+                sds *tokens = sdssplitlen(mpd_host, (ssize_t)strlen(mpd_host), "@", 1, &count);
                 mympd_state->mpd_state->mpd_host = sdsreplace(mympd_state->mpd_state->mpd_host, tokens[1]);
                 mympd_state->mpd_state->mpd_pass = sdsreplace(mympd_state->mpd_state->mpd_pass, tokens[0]);
                 sdsfreesplitres(tokens,count);
@@ -137,7 +137,7 @@ void mympd_autoconf(struct t_mympd_state *mympd_state) {
         }
         const char *mpd_port = getenv("MPD_PORT"); /* Flawfinder: ignore */
         if (mpd_port != NULL && strlen(mpd_port) <= 5) {
-            mympd_state->mpd_state->mpd_port = strtoimax(mpd_port, NULL, 10);
+            mympd_state->mpd_state->mpd_port = (int)strtoimax(mpd_port, NULL, 10);
             MYMPD_LOG_NOTICE("Setting mpd host to \"%d\"", mympd_state->mpd_state->mpd_port);
         }
     }
